refactor: move combobox to submit module
